feat(app): introduce CredentialsApiImpl
Introduces a new class `CredentialsApiImpl` in
`src/utils/credentials-api-impl.ts`, which provides the default
implementation of `CredentialsApi`. Moves the related code from
`src/utils/passquito.ts` to `src/utils/credentials-api-impl.ts`.
